requirements.txt
flask
pycryptodome
pandas
PyPDF2